+++ smartsim/experiment.py
@@ -38,7 +38,8 @@ from .error import SmartSimError
 from .generation import Generator
 from .settings import settings
 from .utils import get_logger
-from .utils.helpers import colorize, detect_launcher, init_default
+from .utils.helpers import colorize, init_default
+from .utils.wlm import detect_launcher
 
 logger = get_logger(__name__)
 
@@ -65,7 +66,7 @@ class Experiment:
         :param exp_path: path to location of ``Experiment`` directory if generated
         :type exp_path: str, optional
         :param launcher: type of launcher being used, options are "slurm", "pbs",
-                         "cobalt", "lsf", or "local", if set to "auto",
+                         "cobalt", "lsf", or "local". If set to "auto",
                          an attempt will be made to find an available launcher on the system.
                          Defaults to "local"
         :type launcher: str, optional
